AiselAssignmentApi/requirements.txt
Django>=4.1.3
djangorestframework>=3.0.0
djangorestframework_simplejwt>=5.0.0
